requirements.txt
django==2.2.13
django-background-tasks==1.2.5
django-cors-headers>=3.2.1
django-filter==2.2.0
djangorestframework>=3.10.3
djangorestframework-camel-case>=1.2.0
dj-database-url==0.5.0
dj-rest-auth==1.0.1
google-api-python-client==1.12.4
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
gunicorn==19.9.0
python-dateutil==2.8.1
requests==2.22.0
sentry-sdk==0.14.4
slackclient>=2.2.0
urllib3==1.25.9
uwsgi==2.0.18
whitenoise==4.1.4

# I had a little trouble installing psycopg2, this finally worked
# brew install postgres
# LDFLAGS='-L/usr/local/lib -L/usr/local/opt/openssl/lib -L/usr/local/opt/readline/lib' pip install psycopg2>=2.8
psycopg2>=2.8
